Allow `Open With' on Windows
On Windows, you are able to open particular file types with particular
applications by double-clicking. This works by sending the path of the
file as a command-line argument to the application. This commit allows
exSCA to automatically open an .esc file sent as a command-line argument,
thus allowing this `open with' functionality on Windows.