Apply changes to already-changed word
The purpose of this commit is probably best illustrated with an example.
Suppose we were executing the change 'VCe/VC@1/_' on the word 'anene'.
exSCA would first search ahead for the first occurrence of 'VCe', which
happens to be 'ane', and replace this with 'ana', forming 'anane'. Next
it will search ahead IN THE ORIGINAL WORD for the next occurrence,
finding 'ene' and replacing this with 'ene'. forming 'anene'. However,
it should have looked in THE ALREADY-CHANGED WORD 'anane', finding 'ane'
and replacing it with 'a', forming 'anana'. This commit makes all changes
apply to the already-changed word, so the previous example will work
correctly.