Add bounds check immediately before `word.at(curIndex)`
Later in the file, we find that every time `word.at(curIndex)` is used,
it is prefixed a line or two before it by a simple bounds check.
However, for some reason, this example has a much more complex
bounds-check before it. which can be replaced with a much simpler check.
This also fixes a bug which causes rules like '/a/_' to not insert at
the end of a word.